+++ backend/src/main/java/com/employed/bar/domain/model/manager/EmployeeSummary.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 @Value
 public class EmployeeSummary {
     String employeeName;
-    double totalHoursWorked;
+    BigDecimal totalHoursWorked;
     BigDecimal totalEarnings;
     BigDecimal totalConsumptions;
     BigDecimal netPay;
